refacto(organizations): move list api internal code
